refactored into folder structure and removed comment
