+++ python/packages/ai/teams/teams_attachment_downloader/__init__.py
@@ -1,0 +1,9 @@
+"""
+Copyright (c) Microsoft Corporation. All rights reserved.
+Licensed under the MIT License.
+"""
+
+from .teams_attachment_downloader import TeamsAttachmentDownloader
+from .teams_attachment_downloader_options import TeamsAttachmentDownloaderOptions
+
+__all__ = ["TeamsAttachmentDownloader", "TeamsAttachmentDownloaderOptions"]
